Add tests for --seed flag.

+++ libchecktestdata.cc
@@ -1256,7 +1256,7 @@ void init_checktestdata(std::istream &progstream, int opt_mask, unsigned long se
 	}
 
 	// Initialize random generator
-	if ( seed == -1) {
+	if ( seed == static_cast<unsigned long>(-1)) {
 		struct timespec time;
 		clock_gettime(CLOCK_REALTIME, &time);
 		mpz_class mpz_seed = 1000000000 * mpz_class(time.tv_sec) + time.tv_nsec;
